refactor: Metric refactored to domain root
