+++ src/main/java/io/pakland/mdas/githubstats/domain/repository/MetricRepository.java
@@ -1,8 +1,7 @@
 package io.pakland.mdas.githubstats.domain.repository;
 
-import io.pakland.mdas.githubstats.domain.entity.Metric;
+import io.pakland.mdas.githubstats.domain.Metric;
 import io.pakland.mdas.githubstats.infrastructure.shell.model.ShellRequest;
-import org.springframework.data.repository.query.Param;
 
 import java.util.List;
