+++ src/main/java/io/pakland/mdas/githubstats/infrastructure/postgres/repository/PostgresMetricRepository.java
@@ -1,6 +1,6 @@
 package io.pakland.mdas.githubstats.infrastructure.postgres.repository;
 
-import io.pakland.mdas.githubstats.domain.entity.Metric;
+import io.pakland.mdas.githubstats.domain.Metric;
 import io.pakland.mdas.githubstats.domain.repository.MetricRepository;
 import io.pakland.mdas.githubstats.infrastructure.shell.model.ShellRequest;
 import org.springframework.data.jpa.repository.JpaRepository;
@@ -9,10 +9,6 @@ import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.time.Instant;
-import java.time.YearMonth;
-import java.time.ZoneId;
-import java.time.temporal.ChronoUnit;
 import java.util.List;
 
 @Repository
